+++ src/Microsoft.Health.Fhir.Shared.Web/Startup.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Microsoft.Health.Fhir.Azure;
 
 namespace Microsoft.Health.Fhir.Web
 {
